requirements.txt
alembic==1.6.1
click==7.1.2
Flask==1.1.2
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.11
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1
cffi==1.14.5
cryptography==3.4.7
gunicorn==20.1.0
pycparser==2.20
